Added header file for the operations
